Merging to solve the conflicts

+++ bigdata.common/src/main/java/org/prosolo/bigdata/common/dal/pojo/EventDailyCount.java
@@ -4,7 +4,7 @@ public class EventDailyCount {
 	
 	private String type;
 	
-	private int count;
+	private long count;
 	
 	private long date;
 	
@@ -16,11 +16,11 @@ public class EventDailyCount {
 		this.type = type;
 	}
 
-	public int getCount() {
+	public long getCount() {
 		return count;
 	}
 
-	public void setCount(int count) {
+	public void setCount(long count) {
 		this.count = count;
 	}
 
@@ -32,7 +32,7 @@ public class EventDailyCount {
 		this.date = date;
 	}
 
-	public EventDailyCount(String type, long date, int count) {
+	public EventDailyCount(String type, long date, long count) {
 		this.type = type;
 		this.date = date;
 		this.count = count;
